# docs/openapi.yaml
openapi: 3.0.3
info:
  title: InvestIQ API
  description: |
    InvestIQ is a comprehensive algorithmic trading platform with multi-engine analysis,
    backtesting, portfolio management, and autonomous trading capabilities.

    ## Authentication
    All endpoints (except `/health` and `/metrics`) require authentication via the `X-API-Key` header or Bearer token.

    ## Live Trading Protection
    Broker write endpoints (POST /api/broker/execute, DELETE /api/broker/positions/:symbol, POST /api/broker/orders/:id/cancel)
    require an additional `X-Live-Trading-Key` header for safety.

    ## Response Format
    All API responses follow a standard wrapper:
    ```json
    {
      "success": true,
      "data": { ... },
      "error": null
    }
    ```
  version: 1.0.0
  contact:
    name: InvestIQ Support
    url: https://github.com/yourusername/invest-iq
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.investiq.io
    description: Production server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/broker/positions/{symbol}:
    delete:
      tags: [Broker]
      summary: Close a position (sell all shares)
      description: Requires X-Live-Trading-Key header
      security:
        - ApiKeyAuth: []
        - LiveTradingAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
      responses:
        '200':
          description: Order submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/broker/orders/{id}:
    get:
      tags: [Broker]
      summary: Get a specific order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/broker/orders/{id}/cancel:
    post:
      tags: [Broker]
      summary: Cancel an order
      description: Requires X-Live-Trading-Key header
      security:
        - ApiKeyAuth: []
        - LiveTradingAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/broker/execute:
    post:
      tags: [Broker]
      summary: Execute a trade
      description: |
        Submit a market order. Requires X-Live-Trading-Key header.
        Performs pre-trade risk checks and circuit breaker validation.
      security:
        - ApiKeyAuth: []
        - LiveTradingAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteTradeRequest'
      responses:
        '200':
          description: Trade executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ============================================================================
  # Backtest
  # ============================================================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Can also use Bearer token in Authorization header.

    LiveTradingAuth:
      type: apiKey
      in: header
      name: X-Live-Trading-Key
      description: Additional key required for live trading write operations.

    AdminAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin-level API key required for privileged operations.
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          type: object
          description: Response data (structure varies by endpoint)
        error:
          type: string
          nullable: true
...
    ExecuteTradeRequest:
      type: object
      required: [symbol, action, shares]
      properties:
        symbol:
          type: string
          example: AAPL
        action:
          type: string
          enum: [buy, sell]
        shares:
          type: number
          example: 10
        confidence:
          type: number
          minimum: 0
          maximum: 1
        notes:
          type: string
        idempotency_key:
          type: string
          description: Optional key for idempotent trades (24h expiry)
...